IGNITE-28819 Move tracing classes used by GridNioServer to ignite-commons
Move Scope, SpanStatus, SpanType, Span, SpanManager, NoopSpan, SpanTags and MTC
to ignite-commons, and add NoopSpanManager (no-op SpanManager). GridNioServer
now uses SpanManager/NoopSpanManager instead of Tracing/NoopTracing, removing its
dependency on the core tracing manager. NoopTracing extends NoopSpanManager.
