+++ modules/commons/src/main/java/org/apache/ignite/internal/processors/tracing/NoopSpan.java
