+++ modules/commons/src/main/java/org/apache/ignite/spi/tracing/Scope.java
